requirements.txt
# Deep Learning packages
tensorflow==2.14.0
tensorflow-datasets==4.9.3
tensorflow-probability==0.21.0

# Configuration management
hydra-core==1.3.2


